requirements.txt
# Основные зависимости
Django~=5.1.1
djangorestframework~=3.15.2
psycopg2-binary==2.9.8
django-cors-headers==3.14.0

# Swagger для документации API
drf-yasg==1.21.5

# Celery и Redis
celery~=5.4.0
redis~=5.2.1

# Telegram API
python-telegram-bot==20.3

# Утилиты
python-decouple==3.8
gunicorn==20.1.0  # Для продакшена

# Тестирование
pytest~=8.2.2
pytest-django==4.5.2
pytest-cov==4.1.0

pip~=24.2
attrs~=23.2.0
pillow~=11.0.0
asyncio~=3.4.3
filelock~=3.13.4
amqp~=5.3.1
vine~=5.1.0
pytz~=2024.1
PyYAML~=6.0.1
typing_extensions~=4.10.0
kombu~=5.4.2
billiard~=4.2.1
requests~=2.32.0
click~=8.1.7
python-dateutil~=2.9.0.post0
asgiref~=3.8.1
sqlparse~=0.5.1
psycopg2~=2.9.10
colorama~=0.4.6
numpy~=1.26.4
certifi~=2024.7.4
urllib3~=2.2.2
wcwidth~=0.2.13
six~=1.16.0
packaging~=23.2
inflection~=0.5.1
uritemplate~=4.1.1
idna~=3.7
prompt_toolkit~=3.0.48
filters~=1.3.2